requirements.txt
google-cloud-vision==3.7.2
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.3.0
flask==3.0.3
pytest==8.2.0
